Restructured WAR Working Directories
Changed the system to look for the unique client id
assigned by mage-editor to each client and using that
create a unique directory in which all the work goes on.
Additionally, genome file processing was added to allow
for unique fasta files to be sent to the server

The structure now works as such -> Upon recieving a post
request, the id is extracted, and a new folder is created
All the files from the /master folder are then copied into
the new folder. From here the servlet works in the working
directory.

- Textfile.java: Added directory copying functionality

- MerlinServlet.java : Added id checking

- Optmage.java : Added id checking, file copying, genome
  file generation

- MageEditor.java : Added new constants to each dir